+++ DelvUI/Interface/Party/PartyFramesBar.cs
@@ -29,10 +29,10 @@ namespace DelvUI.Interface.Party
 
         private LabelHud _nameLabelHud;
         private LabelHud _healthLabelHud;
-        private LabelHud _manaLabelHud;
         private LabelHud _orderLabelHud;
         private LabelHud _raiseLabelHud;
         private LabelHud _invulnLabelHud;
+        private PrimaryResourceHud _manaBarHud;
         private CastbarHud _castbarHud;
         private StatusEffectsListHud _buffsListHud;
         private StatusEffectsListHud _debuffsListHud;
@@ -42,19 +42,7 @@ namespace DelvUI.Interface.Party
 
         private SmoothHPHelper _smoothHPHelper = new SmoothHPHelper();
 
-        private IPartyFramesMember? _member = null;
-        public IPartyFramesMember? Member
-        {
-            get => _member;
-            set
-            {
-                if (_member == value) { return; }
-
-                _member = value;
-                _smoothHPHelper.Reset();
-            }
-        }
-
+        public IPartyFramesMember? Member;
 
         public PartyFramesBar(
             string id,
@@ -81,11 +69,11 @@ namespace DelvUI.Interface.Party
 
             _nameLabelHud = new LabelHud(config.NameLabelConfig);
             _healthLabelHud = new LabelHud(config.HealthLabelConfig);
-            _manaLabelHud = new LabelHud(_manaBarConfig.ValueLabelConfig);
             _orderLabelHud = new LabelHud(config.OrderLabelConfig);
             _raiseLabelHud = new LabelHud(_raiseTrackerConfig.LabelConfig);
             _invulnLabelHud = new LabelHud(_invulnTrackerConfig.LabelConfig);
 
+            _manaBarHud = new PrimaryResourceHud(_manaBarConfig, "");
             _castbarHud = new CastbarHud(_castbarConfig, "");
             _buffsListHud = new StatusEffectsListHud(buffsConfig, "");
             _debuffsListHud = new StatusEffectsListHud(debuffsConfig, "");
@@ -147,18 +135,22 @@ namespace DelvUI.Interface.Party
             {
                 MouseOverHelper.Instance.Target = character;
 
-                // move player bar to this spot on ctrl+alt+shift click
-                if (ImGui.GetIO().KeyCtrl && ImGui.GetIO().KeyAlt && ImGui.GetIO().KeyShift && ImGui.GetIO().MouseClicked[0])
+                // left click
+                if (MouseOverHelper.Instance.LeftButtonClicked)
                 {
-                    MovePlayerEvent?.Invoke(this);
+                    // move player bar to this spot on ctrl+alt+shift click
+                    if (ImGui.GetIO().KeyCtrl && ImGui.GetIO().KeyAlt && ImGui.GetIO().KeyShift)
+                    {
+                        MovePlayerEvent?.Invoke(this);
+                    }
+                    // target
+                    else if (character != null)
+                    {
+                        Plugin.TargetManager.SetTarget(character);
+                    }
                 }
-                // target
-                else if (ImGui.IsMouseClicked(ImGuiMouseButton.Left) && character != null)
-                {
-                    Plugin.TargetManager.SetTarget(character);
-                }
-                // context menu
-                else if (ImGui.IsMouseClicked(ImGuiMouseButton.Right))
+                // right click (context menu)
+                else if (MouseOverHelper.Instance.RightButtonClicked)
                 {
                     OpenContextMenuEvent?.Invoke(this);
                 }
@@ -226,22 +218,6 @@ namespace DelvUI.Interface.Party
             var borderSize = _config.Size + Vector2.One * 2;
             var color = borderColor != null ? borderColor.Base : _config.ColorsConfig.BorderColor.Base;
             drawList.AddRect(borderPos, borderPos + borderSize, color);
-
-            // mana
-            if (ShowMana())
-            {
-                var parentPos = Utils.GetAnchoredPosition(Position, -_config.Size, _manaBarConfig.HealthBarAnchor);
-                var manaBarPos = Utils.GetAnchoredPosition(parentPos + _manaBarConfig.Position, _manaBarConfig.Size, _manaBarConfig.Anchor);
-
-                drawList.AddRectFilled(manaBarPos, manaBarPos + _manaBarConfig.Size, _manaBarConfig.BackgroundColor.Base);
-
-                var scale = (float)Member.MP / (float)Member.MaxMP;
-                var fillSize = new Vector2(Math.Max(1, _config.Size.X * scale), _manaBarConfig.Size.Y);
-
-                DrawHelper.DrawGradientFilledRect(manaBarPos, fillSize, _manaBarConfig.Color, drawList);
-
-                _manaLabelHud.Draw(manaBarPos, _manaBarConfig.Size, character);
-            }
 
             // role/job icon
             if (_roleIconConfig.Enabled && Member.JobId > 0)
@@ -322,21 +298,8 @@ namespace DelvUI.Interface.Party
             if (ShowMana())
             {
                 var parentPos = Utils.GetAnchoredPosition(Position, -_config.Size, _manaBarConfig.HealthBarAnchor);
-                var manaBarPos = Utils.GetAnchoredPosition(parentPos + _manaBarConfig.Position, _manaBarConfig.Size, _manaBarConfig.Anchor);
-
-                if (character == null)
-                {
-                    string oldText = _manaBarConfig.ValueLabelConfig.GetText();
-                    _manaBarConfig.ValueLabelConfig.SetText(Member.MP.ToString());
-
-                    _manaLabelHud.Draw(manaBarPos, _manaBarConfig.Size, character);
-
-                    _manaBarConfig.ValueLabelConfig.SetText(oldText);
-                }
-                else
-                {
-                    _manaLabelHud.Draw(manaBarPos, _manaBarConfig.Size, character);
-                }
+                _manaBarHud.Actor = character;
+                _manaBarHud.Draw(parentPos);
             }
 
             // buffs / debuffs