+++ DelvUI/Interface/Party/PartyFramesInvulnTracker.cs
@@ -52,6 +52,7 @@ namespace DelvUI.Interface.Party
 
                 if (member.Character is not BattleChara battleChara || member.HP <= 0)
                 {
+                    member.InvulnStatus = null;
                     continue;
                 }
 
@@ -60,9 +61,9 @@ namespace DelvUI.Interface.Party
                 if (tankInvuln == null)
                 {
                     member.InvulnStatus = null;
-                    break;
+                    continue;
                 }
-                
+
                 // apply invuln data based on buff
 
                 member.InvulnStatus = new InvulnStatus(InvulnMap[tankInvuln.StatusId], tankInvuln.RemainingTime, tankInvuln.StatusId);